+++ Backend/database/migrations/2025_01_12_072920_add_status_column_to_competition_applications_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('competition_applications')) {
             Schema::table('competition_applications', function (Blueprint $table) {
-            $table->dropColumn('status');
+            if (Schema::hasColumn('competition_applications', 'status')) { $table->dropColumn('status'); }
         });
         }
     }
