+++ Backend/database/migrations/2025_01_16_090730_add_competition_id_column_to_events_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('events')) {
             Schema::table('events', function (Blueprint $table) {
-$table->dropColumn('competition_id');
+if (Schema::hasColumn('events', 'competition_id')) { $table->dropColumn('competition_id'); }
         });
         }
     }
