+++ Backend/database/migrations/2025_01_16_095826_add_idea_path_id_column_to_mentors_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('mentors')) {
             Schema::table('mentors', function (Blueprint $table) {
-$table->dropColumn('track_id');
+if (Schema::hasColumn('mentors', 'track_id')) { $table->dropColumn('track_id'); }
         });
         }
     }
