+++ Backend/database/migrations/2025_01_19_170512_add_application_id_column_to_teams_table.php
@@ -28,7 +28,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('teams')) {
             Schema::table('teams', function (Blueprint $table) {
-$table->dropColumn('application_id');
+if (Schema::hasColumn('teams', 'application_id')) { $table->dropColumn('application_id'); }
         });
         }
     }
