+++ Backend/database/migrations/2025_02_03_102629_remove_competition_id_colum_from_judges_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('judges')) {
             Schema::table('judges', function (Blueprint $table) {
-$table->dropColumn('competition_id');
+if (Schema::hasColumn('judges', 'competition_id')) { $table->dropColumn('competition_id'); }
         });
         }
     }
