+++ Backend/database/migrations/2025_02_04_121715_add_registered_as_team_column_in_competition_applications_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('competition_applications')) {
             Schema::table('competition_applications', function (Blueprint $table) {
-            $table->dropColumn('registered_as_team');
+            if (Schema::hasColumn('competition_applications', 'registered_as_team')) { $table->dropColumn('registered_as_team'); }
         });
         }
     }
