+++ Backend/database/migrations/2025_02_11_105536_add_role_type_column_to_users_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('users')) {
             Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role_type');
+            if (Schema::hasColumn('users', 'role_type')) { $table->dropColumn('role_type'); }
         });
         }
     }
