+++ Backend/database/migrations/2025_02_16_092145_add_slug_column_to_challenges_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('challenges')) {
             Schema::table('challenges', function (Blueprint $table) {
-            $table->dropColumn('slug');
+            if (Schema::hasColumn('challenges', 'slug')) { $table->dropColumn('slug'); }
         });
         }
     }
