+++ Backend/database/migrations/2025_03_03_162327_add_idea_description_column_to_competition_applications_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('competition_applications')) {
             Schema::table('competition_applications', function (Blueprint $table) {
-            $table->dropColumn('idea_description');
+            if (Schema::hasColumn('competition_applications', 'idea_description')) { $table->dropColumn('idea_description'); }
         });
         }
     }
