+++ Backend/database/migrations/2025_05_26_120624_add_visibility_to_winners_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
     {
         if (Schema::hasTable('winners')) {
             Schema::table('winners', function (Blueprint $table) {
-            $table->dropColumn('is_visible');
+            if (Schema::hasColumn('winners', 'is_visible')) { $table->dropColumn('is_visible'); }
         });
         }
     }
